# promql: use faster heap method for topk/bottomk 